% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_open.R
\name{amount_of_neighbours}
\alias{amount_of_neighbours}
\title{Get the Types of Neighbours and Their Counts}
\usage{
amount_of_neighbours(cd_assign)
}
\arguments{
\item{cd_assign}{A data frame with cluster domain assignments.}
}
\description{
This function calculates the types of neighbours and their counts, and generates a plot and CSV file.
}
