% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_analysis.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate Correlation Matrix}
\usage{
calculate_correlation(df, vector)
}
\arguments{
\item{df}{A data frame containing the data to calculate the correlation matrix.}

\item{vector}{A vector of column names to be used in the correlation matrix.}
}
\value{
A correlation matrix.
}
\description{
This function calculates a correlation matrix based on the co-occurrence of elements in a data frame.
}
