% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{calculate_plot_height}
\alias{calculate_plot_height}
\title{Calculate plot height}
\usage{
calculate_plot_height(neighbour_count)
}
\arguments{
\item{neighbour_count}{A data frame with neighbor counts.}
}
\value{
A numeric value for plot height, or NULL if invalid.
}
\description{
This function calculates an appropriate plot height based on the maximum count.
}
\keyword{internal}
