% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{collec.all.neigbour}
\alias{collec.all.neigbour}
\title{Collect all neighbors for a list of proteins}
\usage{
collec.all.neigbour(protein.assembly, basepairs = 300, m = 15, PATH = "data")
}
\arguments{
\item{protein.assembly}{A data frame with protein and assembly information.}

\item{basepairs}{Number of base pairs to consider for neighbors (default is 300).}

\item{m}{Number of neighbors to find (default is 15).}

\item{PATH}{Path where to ncbi_dataset folder is stored (default = data).}
}
\value{
A data frame with all neighbors.
}
\description{
Collect all neighbors for a list of proteins
}
