% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_neighbors.R
\name{collect_all_neighbours}
\alias{collect_all_neighbours}
\title{Collect all neighbors for a list of proteins}
\usage{
collect_all_neighbours(
  protein.assembly,
  basepairs = 300,
  max_neighbors = 15,
  PATH = "data",
  overlap = 50
)
}
\arguments{
\item{protein.assembly}{A data frame with protein and assembly information.}

\item{basepairs}{Number of base pairs to consider for neighbors (default is 300).}

\item{max_neighbors}{Number of neighbors to find (default is 15).}

\item{PATH}{Path where the ncbi_dataset folder is stored (default = data).}
}
\value{
A data frame with all neighbors.
}
\description{
This function identifies neighbors for all proteins in the protein.assembly data frame.
}
