% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_neighbors.R
\name{collect_all_protein_info}
\alias{collect_all_protein_info}
\title{Collect all protein information}
\usage{
collect_all_protein_info(protein.assembly, PATH = "data")
}
\arguments{
\item{protein.assembly}{A data frame with protein and assembly information.}

\item{PATH}{Path where the ncbi_dataset folder is stored (default = data).}
}
\value{
A data frame with all protein information.
}
\description{
This function collects information about all proteins in the protein.assembly data frame.
}
