% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{combine_and_plot}
\alias{combine_and_plot}
\title{Combine and plot neighbors}
\usage{
combine_and_plot(neighbours_data, cd_data, neighbour_types, clade_assign)
}
\arguments{
\item{neighbours_data}{A data frame with neighbors data.}

\item{cd_data}{A data frame with "conserved domain" data.}

\item{neighbour_types}{A data frame with neighbor types.}

\item{clade_assign}{A data frame with clade assignments.}
}
\value{
A combined data frame.
}
\description{
Combine and plot neighbors
}
