% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_neighbors.R
\name{concatenate_columns}
\alias{concatenate_columns}
\title{Concatenate columns in a data frame}
\usage{
concatenate_columns(df, columns_vector, new_column_name = "concatenated")
}
\arguments{
\item{df}{The data frame.}

\item{columns_vector}{Vector of column names to concatenate.}

\item{new_column_name}{Name for the new column.}
}
\value{
Data frame with the new concatenated column.
}
\description{
This utility function concatenates specified columns into a new column.
}
\keyword{internal}
