% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{create_clade_histograms2}
\alias{create_clade_histograms2}
\title{Create Clade Histograms}
\usage{
create_clade_histograms2(
  fasta_data,
  clade_colors = c("#FFD92F", "#A6D854", "#FC8D62", "#E78AC3", "#8DA0CB", "#66C2A5",
    "#56B4E9", "#E5C494", "#B3B3B3", "pink", "green"),
  output_dir = NULL,
  width = 10,
  height = 10
)
}
\arguments{
\item{fasta_data}{A data frame containing data with columns for organism and clade.}

\item{clade_colors}{A vector of colors to be used for the clades. Default is a predefined set of colors.}

\item{output_dir}{Directory to save the plot (default is based on current date).}

\item{width}{Plot width in cm (default is 10).}

\item{height}{Plot height in cm (default is 10).}
}
\value{
A ggplot object representing the combined histograms.
}
\description{
This function creates histograms for each clade showing the distribution of
proteins per organism.
}
