% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{create_clade_labels}
\alias{create_clade_labels}
\title{Create clade labels}
\usage{
create_clade_labels(representatives)
}
\arguments{
\item{representatives}{A data frame with representative counts by clade.}
}
\value{
A named vector with clade labels.
}
\description{
This function creates labeled strings for clades.
}
\keyword{internal}
