% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{create_dummy_annotations}
\alias{create_dummy_annotations}
\title{Create dummy COG annotations for testing}
\usage{
create_dummy_annotations(protein_ids, output_dir)
}
\arguments{
\item{protein_ids}{Vector of protein IDs to annotate}

\item{output_dir}{Directory to save the dummy results}
}
\value{
A data frame with dummy COG annotations
}
\description{
Create dummy COG annotations for testing
}
