% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_neighbors.R
\name{data.frame2fasta}
\alias{data.frame2fasta}
\title{Convert data frame to FASTA format}
\usage{
data.frame2fasta(
  df,
  name_vector,
  sequence_column = "seq",
  output_file = "fasta.fasta"
)
}
\arguments{
\item{df}{The data frame containing sequence data.}

\item{name_vector}{Vector of column names to use for sequence names.}

\item{sequence_column}{Name of the column containing sequences.}

\item{output_file}{Path for the output FASTA file.}
}
\value{
Invisible NULL, called for side effects.
}
\description{
This function creates a FASTA file from a data frame with sequence information.
}
