% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_script}
\alias{find_script}
\title{Get the path to a script file}
\usage{
find_script(script_name, default_locations = c())
}
\arguments{
\item{script_name}{The name of the script file}

\item{default_locations}{Additional locations to check}
}
\value{
The full path to the script file, or NULL if not found
}
\description{
This function finds the path to a script file, trying multiple locations.
}
\keyword{internal}
