% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{generate_color_vector}
\alias{generate_color_vector}
\title{Generate color vector}
\usage{
generate_color_vector(n = 20)
}
\arguments{
\item{n}{The number of colors to generate (default is 20).}
}
\value{
A vector of color codes.
}
\description{
This function generates a color vector for plots.
}
\keyword{internal}
