% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{generate_dummy_sequences}
\alias{generate_dummy_sequences}
\title{Generate dummy sequences for testing}
\usage{
generate_dummy_sequences(protein_ids, output_file)
}
\arguments{
\item{protein_ids}{Vector of protein IDs}

\item{output_file}{Path to save the FASTA file}
}
\value{
TRUE if successful
}
\description{
Generate dummy sequences for testing
}
\keyword{internal}
