% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{get_output_file_path}
\alias{get_output_file_path}
\title{Get Output File Path Based on Date}
\usage{
get_output_file_path(basepairs, max_neighbors)
}
\arguments{
\item{basepairs}{Number of base pairs to consider for neighbors.}

\item{max_neighbors}{Number of neighbors to find.}
}
\value{
The output file path.
}
\description{
This function asks the user for a date, gets the current date, and determines the output file path based on the provided date or the current date.
If no date is supplied, a folder structure for the curent date is created.
}
