% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{getAttributeField}
\alias{getAttributeField}
\title{Extract a specific field from the attribute column of a GFF file}
\usage{
getAttributeField(x, field, attrsep = ";")
}
\arguments{
\item{x}{The attribute column as a character vector.}

\item{field}{The field to extract.}

\item{attrsep}{The separator used in the attribute column (default is ";").}
}
\value{
A character vector with the extracted field values.
}
\description{
Extract a specific field from the attribute column of a GFF file
}
