% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_neighbors.R
\name{getProteinInfoFromGff}
\alias{getProteinInfoFromGff}
\title{Get protein information from GFF file}
\usage{
getProteinInfoFromGff(input, protein.id)
}
\arguments{
\item{input}{The path to the GFF file.}

\item{protein.id}{The protein ID to search for.}
}
\value{
A data frame with the protein information.
}
\description{
This function extracts information about a specific protein from a GFF file.
}
