% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{getProteinNeighborsFromGff3}
\alias{getProteinNeighborsFromGff3}
\title{Get protein neighbors from a GFF3 file}
\usage{
getProteinNeighborsFromGff3(input, protein.id, basepairs = 300, m = 15)
}
\arguments{
\item{input}{The path to the GFF3 file.}

\item{protein.id}{The protein ID to search for neighbors.}

\item{basepairs}{Number of base pairs to consider for neighbors (default is 300).}

\item{m}{Number of neighbors to find (default is 15).}
}
\value{
A data frame with neighboring proteins.
}
\description{
Get protein neighbors from a GFF3 file
}
