% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_analysis.R
\name{how_many_clades_per_assembly}
\alias{how_many_clades_per_assembly}
\title{Calculate the Number of Clades per Organism}
\usage{
how_many_clades_per_assembly(fasta_data)
}
\arguments{
\item{fasta_data}{A data frame containing data with columns for organism and clade.}
}
\value{
A data frame with the number of clades per organism.
}
\description{
This function calculates the number of clades per organism from the provided data.
}
