% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_config}
\alias{load_config}
\title{Load configuration from YAML file}
\usage{
load_config(config_file = "config/config.yaml", override_params = NULL)
}
\arguments{
\item{config_file}{Path to the configuration YAML file}

\item{override_params}{Named list of parameters that override the config file values}
}
\value{
A list containing the configuration parameters
}
\description{
This function loads configuration parameters from a YAML file.
If a configuration file is not provided, it uses the default configuration.
}
