% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{make_correlation_matrix}
\alias{make_correlation_matrix}
\title{Create and Save Correlation Matrix Plot}
\usage{
make_correlation_matrix(
  df,
  vector,
  supress_output = FALSE,
  output_dir = NULL,
  width = 15,
  height = 15
)
}
\arguments{
\item{df}{A data frame containing the data to calculate the correlation matrix.}

\item{vector}{A vector of column names to be used in the correlation matrix.}

\item{supress_output}{Logical indicating whether to suppress the output. Default is FALSE.}

\item{output_dir}{Directory to save the plot (default is based on current date).}

\item{width}{Plot width in cm (default is 10).}

\item{height}{Plot height in cm (default is 10).}
}
\value{
A ggplot object representing the correlation matrix plot if supress_output is FALSE.
}
\description{
This function calculates a correlation matrix, plots it, and saves the plot as a PNG file.
}
