% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{override_config}
\alias{override_config}
\title{Override configuration parameters}
\usage{
override_config(config, override_params)
}
\arguments{
\item{config}{The configuration list}

\item{override_params}{Named list of parameters that override the config}
}
\value{
The modified configuration list
}
\description{
This function overrides the configuration with user-provided parameters.
}
\keyword{internal}
