% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{plot.neighbours}
\alias{plot.neighbours}
\title{Plot neighbors}
\usage{
\method{plot}{neighbours}(all.neighbours.df, protein.id)
}
\arguments{
\item{all.neighbours.df}{A data frame with all neighbors.}

\item{protein.id}{The protein ID to plot.}
}
\description{
Plot neighbors
}
