% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Plot Correlation Matrix}
\usage{
plot_correlation_matrix(correlation.matrix)
}
\arguments{
\item{correlation.matrix}{A correlation matrix to be plotted.}
}
\value{
A ggplot object representing the correlation matrix plot.
}
\description{
This function plots a correlation matrix.
}
