% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_neighbors.R
\name{plot_neighbours}
\alias{plot_neighbours}
\title{Plot neighbors of a specific protein}
\usage{
plot_neighbours(
  all.neighbours.df,
  protein.id,
  output_dir = "output",
  width = 20,
  height = 5
)
}
\arguments{
\item{all.neighbours.df}{A data frame with all neighbors.}

\item{protein.id}{The protein ID to plot.}

\item{output_dir}{Directory to save the plot (default is "output").}

\item{width}{Plot width in cm (default is 20).}

\item{height}{Plot height in cm (default is 5).}
}
\value{
Invisible NULL, called for side effects.
}
\description{
This function plots the genomic neighbors of a specific protein.
}
