% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pn_input_files}
\alias{pn_input_files}
\title{Log input file information for reproducibility}
\usage{
pn_input_files(config)
}
\arguments{
\item{config}{The configuration list}
}
\value{
A data frame with input file information
}
\description{
This function logs information about input files for reproducibility.
}
