% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pn_log_level_priority}
\alias{pn_log_level_priority}
\title{Get the numeric priority of a log level}
\usage{
pn_log_level_priority(level)
}
\arguments{
\item{level}{The log level string}
}
\value{
Numeric priority (higher = more severe)
}
\description{
Get the numeric priority of a log level
}
\keyword{internal}
