% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pn_log_message}
\alias{pn_log_message}
\title{Log a message with a specific level}
\usage{
pn_log_message(level, ...)
}
\arguments{
\item{level}{The log level (DEBUG, INFO, WARN, ERROR, FATAL)}

\item{...}{The message components to log}
}
\description{
Log a message with a specific level
}
\keyword{internal}
