% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pn_setup_logging}
\alias{pn_setup_logging}
\title{Set up standalone logging}
\usage{
pn_setup_logging(config)
}
\arguments{
\item{config}{The configuration list}
}
\value{
The log file path
}
\description{
This function sets up a standalone logging system with no external dependencies
}
