% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{prepare_neighbour_count}
\alias{prepare_neighbour_count}
\title{Prepare neighbor count data}
\usage{
prepare_neighbour_count(
  data,
  exclude_unknown_clade,
  exclude_unknown_cog,
  plot_count_codh = FALSE
)
}
\arguments{
\item{data}{A data frame with combined neighbor data.}

\item{exclude_unknown_clade}{Logical indicating whether to exclude "unknown" clades.}

\item{exclude_unknown_cog}{Logical indicating whether to exclude "unknown" COG_category.}

\item{plot_count_codh}{Logical indicating whether to plot CODH counts.}
}
\value{
A data frame with neighbor count data.
}
\description{
This function prepares the neighbor count data for plotting.
}
\keyword{internal}
