% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{prepare_neighbour_count_per_protein}
\alias{prepare_neighbour_count_per_protein}
\title{Prepare neighbor count per protein data}
\usage{
prepare_neighbour_count_per_protein(
  data,
  exclude_unknown_clade,
  exclude_unknown_cog
)
}
\arguments{
\item{data}{A data frame with combined neighbor data.}

\item{exclude_unknown_clade}{Logical indicating whether to exclude "unknown" clades.}

\item{exclude_unknown_cog}{Logical indicating whether to exclude "unknown" COG_category.}
}
\value{
A data frame with neighbor counts per protein.
}
\description{
This function prepares the neighbor count per protein data.
}
\keyword{internal}
