% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{prepare_representatives}
\alias{prepare_representatives}
\title{Prepare representatives data}
\usage{
prepare_representatives(data, exclude_unknown_clade)
}
\arguments{
\item{data}{A data frame with combined neighbor data.}

\item{exclude_unknown_clade}{Logical indicating whether to exclude "unknown" clades.}
}
\value{
A data frame with representative counts by clade.
}
\description{
This function prepares the representatives data for plotting.
}
\keyword{internal}
