% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{prepare_total_neighbours}
\alias{prepare_total_neighbours}
\title{Prepare total neighbors data}
\usage{
prepare_total_neighbours(data, exclude_unknown_clade)
}
\arguments{
\item{data}{A data frame with combined neighbor data.}

\item{exclude_unknown_clade}{Logical indicating whether to exclude "unknown" clades.}
}
\value{
A data frame with total neighbor counts by clade.
}
\description{
This function prepares the total neighbors data for plotting.
}
\keyword{internal}
