% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_io.R
\name{process_clade_file}
\alias{process_clade_file}
\title{Process clade file}
\usage{
process_clade_file(file, clade)
}
\arguments{
\item{file}{The path to the clade file.}

\item{clade}{The clade identifier.}
}
\value{
A data frame with processed clade information.
}
\description{
This function reads and processes a single clade file.
}
