% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{process_eggnog_output}
\alias{process_eggnog_output}
\title{Process eggNOG-mapper output into COG classifier format}
\usage{
process_eggnog_output(annotations_file, output_dir)
}
\arguments{
\item{annotations_file}{Path to the eggNOG-mapper annotations file}

\item{output_dir}{Directory to save processed files}
}
\value{
A data frame with COG classifications or NULL if failed
}
\description{
Process eggNOG-mapper output into COG classifier format
}
\keyword{internal}
