% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{protein.alias}
\alias{protein.alias}
\title{Find the protein alias}
\usage{
protein.alias(
  protein.id,
  alias = PROTEIN_ALIAS,
  verbose = FALSE,
  identi = FALSE
)
}
\arguments{
\item{protein.id}{The protein ID to search for.}

\item{alias}{The alias database (default is PROTEIN_ALIAS).}

\item{verbose}{Whether to print additional information (default is FALSE).}

\item{identi}{Whether to include identity information (default is FALSE).}
}
\value{
A data frame with the corresponding PIGI and alias.
}
\description{
Find the protein alias
}
