% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_io.R
\name{read_annotations}
\alias{read_annotations}
\title{Read Neighbour Annotations}
\usage{
read_annotations(current_date, output_dir = NULL, empty_cells = "unknown")
}
\arguments{
\item{current_date}{A string representing the current date, used to locate the annotated CSV file.}

\item{output_dir}{Alternative output directory (overrides date-based directory).}

\item{empty_cells}{A string representing the value to replace empty cells with. Default is "unknown".}
}
\value{
A data frame containing the annotated neighbours if the file exists, otherwise NULL.
}
\description{
This function re-imports the types of neighbours after manual annotation.
}
