% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_open.R
\name{read_cluster_domain}
\alias{read_cluster_domain}
\title{Read Cluster Domain Information}
\usage{
read_cluster_domain(
  PATH = "data",
  classification_path = file.path("classification", "COG")
)
}
\arguments{
\item{classification_path}{The path to the classification directory (default is 'data/classification/COG').}
}
\value{
A data frame with cluster domain assignments.
}
\description{
This function reads and processes cluster domain information from text files.
}
