% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_open.R
\name{read_protein_assembly_data}
\alias{read_protein_assembly_data}
\title{Read Protein and Assembly Data}
\usage{
read_protein_assembly_data(
  protein_file = "proteins.txt",
  assembly_file = "assm_accs.txt",
  protein_assembly_file = "assm_accs_proteins.txt",
  PATH = "data"
)
}
\arguments{
\item{protein_file}{The path to the protein file (default is 'data/proteins.txt').}

\item{assembly_file}{The path to the assembly file (default is 'data/assm_accs.txt').}

\item{protein_assembly_file}{The path to the protein assembly file (default is 'data/assm_accs_proteins.txt').}
}
\value{
A list containing the protein, assembly, and protein assembly data frames.
}
\description{
This function reads the protein and assembly data from specified files.
}
