% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_io.R
\name{read_representatives}
\alias{read_representatives}
\title{Read Representatives}
\usage{
read_representatives(
  PATH = "data",
  path = "representatives",
  ipg_file = "ipg_representative.txt",
  pdb_file = "pdb_representative.txt",
  cluster_file = "cluster_representative.txt"
)
}
\arguments{
\item{PATH}{The base path to the data directory.}

\item{path}{The subdirectory containing the alias files.}

\item{ipg_file}{The name of the IPG alias file.}

\item{pdb_file}{The name of the PDB alias file.}

\item{cluster_file}{The name of the cluster alias file.}
}
\value{
A data frame with combined alias data, or an empty data frame if no alias files are found.
}
\description{
This function reads the IPG, PDB, and cluster alias data from text files and combines them.
}
