% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{round_any}
\alias{round_any}
\title{Rounding function}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{The number to be rounded.}

\item{accuracy}{The accuracy to which the number should be rounded.}

\item{f}{The rounding function to use (default is round).}
}
\value{
The rounded number.
}
\description{
Rounding function
}
