% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{run_cog_classifier}
\alias{run_cog_classifier}
\title{Run COG classifier for protein annotation}
\usage{
run_cog_classifier(fasta_file, output_dir, config)
}
\arguments{
\item{fasta_file}{Path to the FASTA file.}

\item{output_dir}{Output directory for results.}

\item{config}{The configuration list.}
}
\value{
A data frame with the annotation results.
}
\description{
This function runs COG classifier on a FASTA file to annotate proteins.
}
\keyword{internal}
