% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plotting.R
\name{save_plot_with_colors}
\alias{save_plot_with_colors}
\title{Save plot with colors}
\usage{
save_plot_with_colors(plot, filename, col_vector, width = 30, height = 15)
}
\arguments{
\item{plot}{The ggplot object to save.}

\item{filename}{The output filename.}

\item{col_vector}{A vector of color codes.}

\item{width}{Plot width in cm (default is 30).}

\item{height}{Plot height in cm (default is 15).}
}
\value{
Invisible NULL, called for side effects.
}
\description{
This function saves a plot with custom colors.
}
\keyword{internal}
