% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_analysis.R
\name{amount_of_neighbours}
\alias{amount_of_neighbours}
\title{Analysis Functions}
\usage{
amount_of_neighbours(cog_data, output_dir = NULL)
}
\arguments{
\item{cog_data}{Data frame with COG information.}

\item{output_dir}{Directory to save output files (default is based on current date).}
}
\value{
A data frame with types of neighbors and their counts.
}
\description{
This file contains functions for analyzing protein neighbors,
combining data, and calculating statistics.
}
\author{
Maximilian Böhm
}
