% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_analysis.R
\name{combine_and_plot}
\alias{combine_and_plot}
\title{Combine and plot neighbors}
\usage{
combine_and_plot(
  neighbours_data,
  cog_data,
  clade_assign,
  neighbour_annotations,
  output_dir = NULL
)
}
\arguments{
\item{neighbours_data}{A data frame with neighbors data.}

\item{cog_data}{A data frame with COG information from analyze_proteins function.}

\item{clade_assign}{A data frame with clade assignments.}

\item{neighbour_annotations}{A data frame with manual neighbor annotations.}

\item{output_dir}{Directory to save output files (default is based on current date).}
}
\value{
A combined data frame.
}
\description{
This function combines data from different sources and prepares it for visualization.
}
