% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_neighbors.R
\name{getAttributeField}
\alias{getAttributeField}
\title{Neighbor Identification Functions}
\usage{
getAttributeField(x, field, attrsep = ";")
}
\arguments{
\item{x}{The attribute column as a character vector.}

\item{field}{The field to extract.}

\item{attrsep}{The separator used in the attribute column (default is ";").}
}
\value{
A character vector with the extracted field values.
}
\description{
This file contains functions for identifying neighboring proteins
in genomic data, parsing GFF files, and extracting attributes.
}
\author{
Maximilian Böhm
}
