% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_functions.R
\name{plot_neighbours_per_clade}
\alias{plot_neighbours_per_clade}
\title{Plot neighbors per clade}
\usage{
plot_neighbours_per_clade(combined_data)
}
\arguments{
\item{combined_data}{A combined data frame with neighbors and clade information.}
}
\description{
Plot neighbors per clade
}
