% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_open.R
\name{read_clades}
\alias{read_clades}
\title{Read clades}
\usage{
read_clades(
  PATH = "data",
  clade_dir = "clades",
  pattern = "Clade.*_20231130.txt"
)
}
\arguments{
\item{clade_dir}{The directory containing clade files.}

\item{pattern}{The pattern to match clade files.}
}
\value{
A data frame with clade assignments.
}
\description{
This function reads clade information from text files and processes them.
}
